store_manager/requirements.txt
requests
python-dotenv
python-dateutil
Django==4.2.16
gunicorn==23.0.0
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.10
whitenoise==6.6.0
asgiref==3.9.0 
packaging==25.0 
sqlparse==0.5.3
tzdata==2025.2
django-storages==1.14.2
boto3==1.34.84